app/requirements.txt
Flask==0.12.2
requests==2.18.4
lxml==3.8.0
regex==2018.2.21
python-dateutil==2.6.1
numpy==1.14.1
Flask-WTF==0.14.2
WTForms==2.2.1
gspread==0.6.2
oauth2client==4.1.2
Flask-Bootstrap==3.3.7.1
robobrowser==0.5.3
Jinja2==2.10
itsdangerous==0.24
wheel==0.24.0
MarkupSafe==0.23
Werkzeug==0.11.2

